Move FileLock out of repo_utils
